style: nav 관련 스타일 nav로 이동
