+++ client/src/pages/MainPage/Nav/Nav.tsx
@@ -4,7 +4,7 @@ import {Button, Text, Icon, TopNav, IconButton} from '@HDesign/index';
 
 import {ROUTER_URLS} from '@constants/routerUrls';
 
-import {navStyle} from './Nav.style';
+import {navFixedStyle, navStyle, navWrapperStyle} from './Nav.style';
 
 type NavProps = {
   trackStartCreateEvent: () => void;
@@ -19,21 +19,25 @@ const Nav = ({trackStartCreateEvent}: NavProps) => {
   };
 
   return (
-    <header css={navStyle}>
-      <TopNav>
-        <TopNav.Item routePath="/">
-          <IconButton variants="none" aria-label="행동대장 로고">
-            <Icon iconType="heundeut" />
-          </IconButton>
-        </TopNav.Item>
-        <TopNav.Item routePath="/">
-          <Text size="subTitle">행동대장</Text>
-        </TopNav.Item>
-      </TopNav>
-      <Button size="medium" variants="tertiary" onClick={StartCreateEvent} style={{marginRight: '1rem'}}>
-        정산 시작하기
-      </Button>
-    </header>
+    <div css={navFixedStyle}>
+      <div css={navWrapperStyle}>
+        <header css={navStyle}>
+          <TopNav>
+            <TopNav.Item routePath="/">
+              <IconButton variants="none" aria-label="행동대장 로고">
+                <Icon iconType="heundeut" />
+              </IconButton>
+            </TopNav.Item>
+            <TopNav.Item routePath="/">
+              <Text size="subTitle">행동대장</Text>
+            </TopNav.Item>
+          </TopNav>
+          <Button size="medium" variants="tertiary" onClick={StartCreateEvent} style={{marginRight: '1rem'}}>
+            정산 시작하기
+          </Button>
+        </header>
+      </div>
+    </div>
   );
 };
