+++ client/src/pages/MainPage/Nav/index.ts
@@ -1,0 +1,1 @@
+export {default as Nav} from './Nav';
